+++ src/backend/core/admin.py
@@ -1,12 +1,9 @@
 """Admin classes and registrations for core app."""
 
-from django.contrib import admin
 from django.contrib.auth import admin as auth_admin
-from django.contrib.auth import get_user_model
 from django.utils.translation import gettext_lazy as _
 
 
-@admin.register(get_user_model())
 class UserAdmin(auth_admin.UserAdmin):
     """Admin class for the User model"""