+++ src/backend/core/models.py
@@ -7,13 +7,8 @@ from logging import getLogger
 
 from django.conf import settings
 from django.contrib.auth import models as auth_models
-from django.contrib.auth.base_user import AbstractBaseUser
 from django.db import models
 from django.utils.translation import gettext_lazy as _
-
-from timezone_field import TimeZoneField
-
-from core.validators import sub_validator
 
 logger = getLogger(__name__)
 
@@ -92,81 +87,3 @@ class UserManager(auth_models.UserManager):
                     )
                 ) from err
         return None
-
-
-class User(AbstractBaseUser, BaseModel, auth_models.PermissionsMixin):
-    """User model to work with OIDC only authentication."""
-
-    sub = models.CharField(
-        _("sub"),
-        help_text=_("Required. 255 characters or fewer. ASCII characters only."),
-        max_length=255,
-        validators=[sub_validator],
-        unique=True,
-        blank=True,
-        null=True,
-    )
-
-    full_name = models.CharField(_("full name"), max_length=100, null=True, blank=True)
-    short_name = models.CharField(
-        _("short name"), max_length=100, null=True, blank=True
-    )
-
-    email = models.EmailField(_("identity email address"), blank=True, null=True)
-
-    # Unlike the "email" field which stores the email coming from the OIDC token, this field
-    # stores the email used by staff users to login to the admin site
-    admin_email = models.EmailField(
-        _("admin email address"), unique=True, blank=True, null=True
-    )
-
-    language = models.CharField(
-        max_length=10,
-        choices=settings.LANGUAGES,
-        default=None,
-        verbose_name=_("language"),
-        help_text=_("The language in which the user wants to see the interface."),
-        null=True,
-        blank=True,
-    )
-    timezone = TimeZoneField(
-        choices_display="WITH_GMT_OFFSET",
-        use_pytz=False,
-        default=settings.TIME_ZONE,
-        help_text=_("The timezone in which the user wants to see times."),
-    )
-    is_device = models.BooleanField(
-        _("device"),
-        default=False,
-        help_text=_("Whether the user is a device or a real user."),
-    )
-    is_staff = models.BooleanField(
-        _("staff status"),
-        default=False,
-        help_text=_("Whether the user can log into this admin site."),
-    )
-    is_active = models.BooleanField(
-        _("active"),
-        default=True,
-        help_text=_(
-            "Whether this user should be treated as active. "
-            "Unselect this instead of deleting accounts."
-        ),
-    )
-
-    metadata = models.JSONField(
-        default=dict, blank=True, help_text=_("Additional information about the user.")
-    )
-
-    objects = UserManager()
-
-    USERNAME_FIELD = "admin_email"
-    REQUIRED_FIELDS = []
-
-    class Meta:
-        db_table = "accounts_user"
-        verbose_name = _("user")
-        verbose_name_plural = _("users")
-
-    def __str__(self):
-        return self.email or self.admin_email or str(self.id)