+++ src/backend/e2e/viewsets.py
@@ -1,6 +1,7 @@
 """Viewsets for the e2e app."""
 
 from django.contrib.auth import get_user_model, login
+from django.contrib.auth.hashers import make_password
 
 import rest_framework as drf
 from rest_framework import response as drf_response
@@ -27,15 +28,10 @@ class UserAuthViewSet(drf.viewsets.ViewSet):
         serializer = E2EAuthSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
 
-        user = (
-            get_user_model()
-            .objects.filter(email=serializer.validated_data["email"])
-            .first()
+        user, _created = get_user_model().objects.get_or_create(
+            email=serializer.validated_data["email"].lower(),
+            defaults={"password": make_password(None)},
         )
-        if not user:
-            user = get_user_model()(email=serializer.validated_data["email"])
-            user.set_unusable_password()
-            user.save()
 
         login(request, user, "django.contrib.auth.backends.ModelBackend")