+++ CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to
 
 ## [Unreleased]
 
-- ✨(frontend) init the frontend app (webaoo + e2e)
+### Added
 
+- ✨(frontend) init the frontend app (webapp + e2e)
+
+### Changed
+
+- ✨(users) make `email` our username
 
 [unreleased]: https://github.com/suitenumerique/docs/compare/main